인증인가 연결 (#7)
* feat: Controller에서 인증객체를 받아오도록 수정

* feat: ReplyController에 인증객체 받아오도록 수정

* feat: PostService, CommentService, ReplyService 인터페이스에서 유저받아오도록 create, update, delete메서드 수정

* feat: Serivce구현체에 유저 받아올 수 있도록 수정 및 해당부분 반영하여 테스트 수정

- ReplyServiceDecorator
- ReplyServiceImpl
- ReplyServiceImplTest
- CommentServiceDecorator
- CommentServiceImpl
- CommentServiceImplTest

* feat: PostEntity, CommentEntity, ReplyEntity 와 UserEntity 연관관계 설정

* feat: 인증관계를 추가하며 바뀐 PostEntity를 반영하여 Post에 작성자 추가

* feat: PostService의 구현체들에 User를 받아올 수 있도록 수정 / Post 생성시 글쓴이 추가

* feat: PostServiceImplTest에 유저추가하여 수정 ( create의 경우 authorId에 넣을수 있게 추가. )

* feat: PostRepositoryImpl에 도메인변환, 엔티티변환 메서드 수정

* feat: Comment 생성 시 작성자와의 연관관계 반영 완료

* fix: 가입되지 않은 유저가 로그인하면 예외가 아니라 null을 리턴하도록 수정

* feat: 대댓글 생성시 인증객체 정보 확인 및 연결

* fix: dev병합이후 authorId를 userId로 변경

* feat: 권한 검증 관련 에러코드추가

* feat: Post작성자와 요청유저 일치여부 확인 메서드 작성

* feat: 서비스 구현체에서 요청유저와 게시글작성자 일치여부 확인

* feat: Comment 수정 및 삭제시 작성자와 요청유저 일치여부 검증하여 작동하도록 기능 추가

* fix: Comment 도메인 객체 생성시 작성자 id 포함하도록 수정

* feat: Reply 수정, 삭제시 유저의 권한 검증 로직 추가

* chore: 불필요한 열거형 상수 제거

- USER_NOT_FOUND는 UserErrorCode에 이미 존재

* chore: AuthErrorCode 메시지 변경

* refactor: USER를 빌더 패턴으로 수정

* chore: 쓰이지 않는 에러코드 삭제

---------

+++ core/src/main/java/io/devarium/core/domain/reply/service/ReplyServiceImpl.java
@@ -5,6 +5,7 @@ import io.devarium.core.domain.reply.exception.ReplyErrorCode;
 import io.devarium.core.domain.reply.exception.ReplyException;
 import io.devarium.core.domain.reply.port.UpsertReply;
 import io.devarium.core.domain.reply.repository.ReplyRepository;
+import io.devarium.core.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -15,10 +16,11 @@ public class ReplyServiceImpl implements ReplyService {
     private final ReplyRepository replyRepository;
 
     @Override
-    public Reply createReply(UpsertReply request) {
+    public Reply createReply(UpsertReply request, User user) {
         Reply reply = Reply.builder()
             .content(request.content())
             .commentId(request.commentId())
+            .userId(user.getId())
             .build();
         return replyRepository.save(reply);
     }
@@ -35,14 +37,17 @@ public class ReplyServiceImpl implements ReplyService {
     }
 
     @Override
-    public Reply updateReply(Long replyId, UpsertReply request) {
+    public Reply updateReply(Long replyId, UpsertReply request, User user) {
         Reply reply = getReply(replyId);
+        reply.validateAuthor(user.getId());
         reply.updateContent(request.content());
         return replyRepository.save(reply);
     }
 
     @Override
-    public void deleteReply(Long replyId) {
+    public void deleteReply(Long replyId, User user) {
+        Reply reply = getReply(replyId);
+        reply.validateAuthor(user.getId());
         replyRepository.deleteById(replyId);
     }
